base code for ARA::COM Camera Hazard warning
Build system is based on both cmake and qmake for now.

There is a dummy imagesource implemented that simply feeds
an image once a second to QML based images that are available
in qrc.

There is also an Yocto layer included (meta-franca-ara) which
includes an recipe for the application. This is added to simplify
integration with GDP as this repository is private.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>
Signed-off-by: Gunnar Andersson <gandersson@genivi.org>