interface: Add .fdepl files for interfaces
Minimal content to get the SOME/IP generator to even run. I mostly
mirrored another example, and looked into the deployment *specification*
of the SOME/IP generator for some hints.

There are dummy values inserted for most of the deployment details. Some
of them are not strictly necessary to make the generator run since they
were optional, but the generator still warns if no deployment info is
provided at all

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>