Add src-gen files to main project & rm src-gen.pro
Instead of setting src-gen as a project subdirectory, the files are
added to the main project.  If src-gen is set as a subdir then qmake
wants to configure it as a separate entity (lib, app, plugin).
I think a separate .so is not necessary for this case- let's just
link the glue code directly into the main executable.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>