TEMP: Use host-installed CommonAPI & vSomeIP
This is likely temporary.  In the Yocto setup it could be different
but I appreciate the ability to do some early development on native host
with at least successful compilation.

I have simply compiled and installed the libs & headers on my work
machine.  The default install path was /usr/local/* so it needs to be
added to search paths.

If we align the paths it should work both for host and Yocto/SDK setup?

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>