Implement network layer and separate from graphics
1. Initial implementation of callbacks and processing from updates
coming from the CommonAPI C++ framework.  This compiles, so in some
sense it fits together, but I think it's quite far from "communicating"
yet...

This proposes also a rewrite that takes the thread away from
ImageSource and creates a networking thread in a separate class.

Rough design:

The networking layer deals with the CommonAPI layer entirely and
calls functions in ImageSource to trigger updates.  ImageSource
implements (as before) the Qt-style objects that represent the upddates
and communicates them to the QML layer using signals.  ImageSource just
does not have its own thread anymore (its functions are executed from
the network thread).

From a runtime point of view, nothing changed - there is an equal amount
of threads.  From a static/structural point of view there is strict
separation between the datatypes used on CommonAPI (standard C++ only)
and the Qt World, with the boundary between SomeIpNetworkThread class
and ImageSource class.

This division causes some extra layers and boiler-plate code as a
consequence.  We'll see how we like it.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>