vSOMEIP-only test server and client
As the summary indicates, this is programmed directly against vsomeip,
without any CommonAPI bindings.

This really helps testing.  First of all the SOME-IP configuration file
can be tested with a plain vSomeIP server _and_ client that is
independent of the main program.  This client does nothing more than
check that the service is provided in the network, which is a good first
step.

Secondly, the main program (client) will now be possible to connect
including its CommonAPI binding, against this local server.
As of now we check only that there is a "connection" to the requested
service, and later on the server code can simulate more of the ARA
system role to test the IVI part independently.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>