conf: Add a separate application for test client
If the vsomeip/client has the same ID as the capi_client (or the main
application) then there will be complaints.  In theory we should be able
to have more than one client of course, even subscribing to the same
data. So I added a separate application ID for the vsomeip/client

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>