dependencies: Add CommonAPI Build Helper (submodule)
To simplify the current example/test code I will need to regenerate code
again, and to be able to script that it's easier to have a fixed
location of this module, which gives a fixed location of the code
generators (and if needed, runtime libraries).

The dependencies dir is only partly needed because Yocto takes care of
that but I still find it useful to do the initial development natively
so I decided to add it.  But this change means we can drop the runtimes
submodules here, since they're now provided by the script.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>