capi_client: Added (new test program)
This is intended to be a more complete CommonAPI based test client
than simple_capi_client.

In other this one uses the full FIDL interface and mirrors the main
graphical program, but since it is test program it runs only some simple
command-line functionality (printouts).

We use it to do the system integration on network level, until it is
debugged.  That is simpler than with the full graphical application.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>