capi_server: Add new test program
This is a CommonAPI based server test loop.  It can act as the
simulation of the service so that the client program(s) can be tested
without the ARA::COM part present.

It is thus similar to vsomeip/server except that
1. vsomeip/server was just testing a very simple interface and never
built out to the full interface.
2. By using CommonAPI we get the same environment, datatypes, etc. and
the data encoding details have already been confirmed in previous
integration workshop that it shall be according to what CommonAPI
provides, so this is easier to get right than the plain vsomeip/server.